# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: YKS Soru Takibi API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List all questions with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [TYT, AYT, Geometri]
        - name: source
          in: query
          schema:
            type: string
            enum: [Deneme, Banka, Fasikül, Ders Kitabı]
        - name: lesson
          in: query
          schema:
            type: string
        - name: publisher
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [Cozulmedi, DogruCozuldu, YanlisHocayaSor]
        - name: topic
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: isOsymBadge
          in: query
          schema:
            type: boolean
        - name: isPremiumBadge
          in: query
          schema:
            type: boolean
        - name: resourceId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create a new question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionInput"
      responses:
        "201":
          description: Created question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/{id}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get a question by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Not found
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Update a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuestionInput"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /questions/image:
    post:
      operationId: uploadQuestionImage
      tags: [questions]
      summary: Upload a question image (base64)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                imageData:
                  type: string
                mimeType:
                  type: string
              required:
                - imageData
                - mimeType
      responses:
        "200":
          description: Image URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                required:
                  - url
  /questions/review/feed:
    get:
      operationId: getQuestionReviewFeed
      tags: [question-review]
      summary: Get a spaced repetition question feed
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: excludeIds
          in: query
          schema:
            type: string
          description: Comma-separated question IDs to avoid in this batch
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 16
      responses:
        "200":
          description: Question review feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionReviewFeedResponse"
  /questions/review/serve/{id}:
    post:
      operationId: markQuestionReviewServed
      tags: [question-review]
      summary: Mark a question as shown in the review feed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Served state stored
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  deduped:
                    type: boolean
                  nextEligibleAt:
                    type: string
                    nullable: true
                required:
                  - ok
  /questions/review/feedback/{id}:
    post:
      operationId: submitQuestionReviewFeedback
      tags: [question-review]
      summary: Store review feedback for a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionReviewFeedbackInput"
      responses:
        "200":
          description: Feedback stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionReviewFeedbackResponse"
  /questions/{id}/drawing:
    get:
      operationId: getDrawing
      tags: [drawings]
      summary: Get drawing data for a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Drawing data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drawing"
    put:
      operationId: saveDrawing
      tags: [drawings]
      summary: Save drawing data for a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveDrawingInput"
      responses:
        "200":
          description: Saved drawing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drawing"
...
components:
  schemas:
...
    Question:
      type: object
      properties:
        id:
          type: integer
        resourceId:
          type: integer
          nullable: true
        imageUrl:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        lesson:
          type: string
        topic:
          type: string
          nullable: true
        publisher:
          type: string
          nullable: true
        testName:
          type: string
          nullable: true
        testNo:
          type: string
          nullable: true
        choice:
          type: string
          nullable: true
          enum: [A, B, C, D, E, null]
        options:
          type: array
          nullable: true
          items:
            type: object
            properties:
              label:
                type: string
              text:
                type: string
            required:
              - label
              - text
        solutionUrl:
          type: string
          nullable: true
        solutionYoutubeUrl:
          type: string
          nullable: true
        solutionYoutubeStartSecond:
          type: integer
          nullable: true
        solutionYoutubeEndSecond:
          type: integer
          nullable: true
        category:
          type: string
          enum: [TYT, AYT, Geometri]
        source:
          type: string
          enum: [Deneme, Banka, Fasikül, Ders Kitabı]
        status:
          type: string
          enum: [Cozulmedi, DogruCozuldu, YanlisHocayaSor]
        hasDrawing:
          type: boolean
        isOsymBadge:
          type: boolean
        isPremiumBadge:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - lesson
        - category
        - source
        - status
        - hasDrawing
        - isOsymBadge
        - isPremiumBadge
        - createdAt
        - updatedAt
...
    QuestionReviewFeedResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuestionReviewItem"
        pagination:
          type: object
          properties:
            total:
              type: integer
            limit:
              type: integer
            offset:
              type: integer
            hasMore:
              type: boolean
          required:
            - total
            - limit
            - offset
            - hasMore
        algorithm:
          type: object
          properties:
            name:
              type: string
            description:
              type: string
          required:
            - name
            - description
      required:
        - items
        - pagination
    QuestionReviewFeedbackInput:
      type: object
      properties:
        feedback:
          type: string
          enum: [again, wrong, hard, correct, easy, less_often, more_often]
      required:
        - feedback
...
    CreateQuestionInput:
      type: object
      properties:
        resourceId:
          type: integer
          nullable: true
        imageUrl:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        lesson:
          type: string
        topic:
          type: string
          nullable: true
        publisher:
          type: string
          nullable: true
        testName:
          type: string
          nullable: true
        testNo:
          type: string
          nullable: true
        choice:
          type: string
          nullable: true
          enum: [A, B, C, D, E, null]
        options:
          type: array
          nullable: true
          items:
            type: object
            properties:
              label:
                type: string
              text:
                type: string
            required:
              - label
              - text
        solutionUrl:
          type: string
          nullable: true
        solutionYoutubeUrl:
          type: string
          nullable: true
        solutionYoutubeStartSecond:
          type: integer
          nullable: true
        solutionYoutubeEndSecond:
          type: integer
          nullable: true
        category:
          type: string
          enum: [TYT, AYT, Geometri]
        source:
          type: string
          enum: [Deneme, Banka, Fasikül, Ders Kitabı]
        status:
          type: string
          enum: [Cozulmedi, DogruCozuldu, YanlisHocayaSor]
        isOsymBadge:
          type: boolean
        isPremiumBadge:
          type: boolean
      required:
        - lesson
        - topic
        - category
        - source
...
    SaveDrawingInput:
      type: object
      properties:
        canvasData:
          type: string
      required:
        - canvasData
...